requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
openai>=1.0.0
